requirements.txt
discord
asyncio
python-dotenv
psycopg2